pyproject.toml
[project]
name = "piano-midi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.9.2",
    "mido>=1.3.2",
    "mypy>=1.11.2",
    "opencv-contrib-python>=4.10.0.84",
    "opencv-python>=4.10.0.84",
    "opencv-stubs>=0.0.10",
    "pydantic>=2.9.0",
    "pyyaml>=6.0.2",
    "ruff>=0.6.4",
    "typer>=0.12.5",
]

# https://docs.astral.sh/ruff/rules/
[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D", # docstrings we don't really care about for now
    "FIX002", # TODO's are OK for now
    "E501", # longer comments are okay
    "TD", # no specific rules for TODO's
    "PT011", # ValueErrors are OK for now
    "FA102", # Only an issue for older python version
    "PLR0913", # Too many arguments in function is not an issue for now
    "T201", # Prints are not OK, but there is a lot of work to do on a debugging mechanism
    "RET504", # it is actually nice to put breakpoints after assignments before returns
    "BLE001", # general exceptions are accepted for now
    "COM812", "ISC001" # Ruff recommends disabling these rules
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
